Wire up set_filesystem_configuration with modular file system
This PR tries to wire up set_filesystem_configuration with
modular file system so that it is possible to set up the file
system configuration through modular file system plugins.

While set_filesystem_configuration API has been defined in
the past, the API was not wired up with modular file system
so it was not possible to use it from plugins.

This PR adds the missing implementation.

This PR also made several changes:
- values associated with `type_tag` was not defined in the past. This
  PR adds the definition.
- TF_Filesystem_Option_Value definition has been modified as previous
  unnamed union was difficult to assign value.
- set_filesystem_configuration_option was removed as it is almost
  a duplicate of set_filesystem_configuration, and it does not
  match the description.

While the API was modified, since set_filesystem_configuration
was not wired with modular file system, it is not used anyway.
So this will not break user or plugins.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>