ParallelDevice: Disable implicit copy-on
Copy-on isn't impossible, although there are a lot of subtleties if we wanted to do it well (e.g. gradients) in addition to performance issues. It seems reasonably usable without implicit copy-on, so it makes sense to trial without it for now.

Some fiddling with ImmediateExecutionTensorHandle::DebugString to make it more useful for an error message. Removes the TensorHandle override since it's much less useful.

PiperOrigin-RevId: 379496797
Change-Id: If71e29de376ff57b69c2961fb520f11cc73e1f99