pfor: Fix the error message for unsupported list ops
Supporting vectorized variants in the fallback converter would be pretty slow or complicated, since we don't know which part of the TensorList/etc. is being accessed, so we'd need to iterate over the whole thing and split it apart into separate variants. Ops taking vectorized variants should mostly be implemented anyway.

PiperOrigin-RevId: 380262037
Change-Id: I5c46a8b7343a6687520c952ed89d8a08123c45f9