Make the ReorderElementwiseAndMoveOperations pattern safe on quantized ops.
Previously, the pattern may produce invalid data movement ops (e.g. reshape) whose input and output quantization parameters are different.

Note that this is not an issue in TFLite converter, where we only run the pass before quantizer, but it could be an issue if we want to reuse the pass to optimize TFlite models outside TFLite converter.

PiperOrigin-RevId: 380288845
Change-Id: I65cfe5307482996eb556953abf9020f69b9bcea1