[XLA:PJRT][XLA:Python] Add TPU executable de/serialization methods.
The methods are added to PjRtClient and not PjRtExecutable since only
the client has subclasses for each platform, and we need to use
platform-specific functionality.

This is currently only plumbed through for the StreamExecutor + TPU
path. GPU, CPU, and TFRT require lower-level serde methods before we
can add them to the PJRT + Python layers.

This change also:
* Factors out some PjRtStreamExecutorClient::Compile logic into a new
  GetExecutableExtras method, so it can be re-used by the new
  deserialization method.

* Makes PjRtTpuClient non-anonymous, so it can be a friend class to
  PjRtStreamExecutorExecutable in order to call SetUpDonation when
  creating a new executable during deserialization.

PiperOrigin-RevId: 382418203
Change-Id: Ic576a39cbd9e193c5068c77faf3fe932981d89aa