[XLA:CPU/GPU] Fix BF16 conversion/reduce-precision for NaNs with weird payloads
EmitReducePrecisionIR is used for two purposes:
 - kReducePrecision
 - kConvert from F32 -> BF16

The kConvert case takes the result of EmitReducePrecisionIR and shifts it right by 16.
If the F32 was a NaN encoded with a payload only in the low bits, the shift would erase the payload turning the NaN into an infinity.

In this case, we force a NaN payload by setting the qNaN bit.

The kReducePrecision wants the NaN back unchanged.

PiperOrigin-RevId: 383745625
Change-Id: I76a2aeff51d4c37127faf9f9fc51237208709911