Feed input tensors only if shape inference requires constant input
Currently, all constants are fed in the InferenceContext regardless of those are required for shape inference or not. This is wasteful and not even feasible for cases involving very large splat constants. It is not possible to create tensors requiring more memory than what is available on the host.

This may require multiple invocation of shape inference functions but this shouldn't add significant overhead as these functions are generally trivial.

PiperOrigin-RevId: 383784459
Change-Id: I63b04617e202656c4bf541ef7a2eb13d23f02910