Support collectives in RemoteCall
CollectiveExecutor was only available if collective graph key is set. TF2 in combination with collective v2 doesn't really have the concept of collective graph key. In the past a workaround was added to set the graph key to a magic number of v2 ops are present. This is not very maintainable as we are adding more collective ops. It also doesn't cover the case of functional ops.

This changes makes CollectiveExecutor always available.

PiperOrigin-RevId: 384387190
Change-Id: Idc94d5d966465b039dd728c99295993a8df6602b