dockerfiles: pin jedi to allow ipython tab completion
Tab completion doesn't work on Jupyter because of an incompatibility
issue. Using the web notebook, only an obscure error is printed to console
but in the case of ipython console, the whole REPL crashes.

This works around the issue by pinning the jedi library to the last
compatible version.

See: https://github.com/ipython/ipython/issues/12740

Signed-off-by: japm48 <japm48gh@gmail.com>