[tf.data] Clean up for tf.data performance model metric collection.
This CL removes previous support for periodically exporting the performance model information to a file and improves the tfstreamz/ integration to avoid the need to maintain static state.

In addition, this CL extends gauges with support for int64 / string / bool futures -- that is, making it possible to compute the values at the time they are requested.

PiperOrigin-RevId: 384812318
Change-Id: I3f82fc8f4b1ec5a1e0d3ac2a6d45b872f0ad815b