Adds automatic ref-counting (on the underlying resource) to ResourceHandle.
This CL adds an option for ResourceHandle to hold on to a pointer to a resource, in which case the ResourceHandle will call `Ref` on the resource when the handle is copied, and call `Unref` on the resource when the handle is destroyed. This feature will serve as an automatic life-cycle management system for resources.

Resources under this life-cycle management system will bypass ResourceMgr.

This CL doesn't opt in any existing resources to this new life-cycle management system. Follow-up CLs will opt them in gradually.

PiperOrigin-RevId: 385665188
Change-Id: Ic15367a2b860f0cc768bca762590036a746e30a0