Adds resource_creator_scope to strategies; For TPUStrategy-Place one lookup table per worker automatically if initialization is called under scope().
In the future if we would like to define a customized creator for any resource that subclasses CachableResource for any specific Strategy, we can do that by implementing SomeStrategy.extended._resource_creator_scope. For now, we implemented it for StaticHashTable (the type of initializable lookup table that KPL uses) and TPUStrategy.

PiperOrigin-RevId: 386187335
Change-Id: Ifedf9e8bac34b264749c9e3e43636e397b3da73d