[TF:TRT] Remove the logic that infers tensor dynamic ranges.
Previously, when calibration is not used and INT8 precision is requested,
TF-TRT tries to infer the dynamic ranges for tensors. This was required for
older TensorRT versions but no longer required for TensorRT 7+. Removing the
logic that infers tensor dynamic ranges can also avoid quantizing nodes that
users do not want to quantize.

PiperOrigin-RevId: 388230513
Change-Id: I0113f3b5ff43143c66ce75c9f80c4e115b7ce5b0